+++ src/compas_timber/connections/__init__.py
@@ -43,7 +43,6 @@ Exceptions
 
 """
 from .joint import Joint
-from .joint import JointOptions
 from .joint import beam_side_incidence
 from .joint import BeamJoinningError
 from .t_butt import TButtJoint
@@ -58,7 +57,6 @@ from .solver import find_neighboring_beams
 
 __all__ = [
     "Joint",
-    "JointOptions",
     "beam_side_incidence",
     "BeamJoinningError",
     "TButtJoint",
