+++ src/compas_timber/connections/joint.py
@@ -156,36 +156,4 @@ class Joint(Data):
         return self._ends
 
 
-class JointOptions(object):
-    """Container for options to be passed to a joint.
 
-    This allows delaying the actual joining of the beams to a downstream component.
-
-    Parameters
-    ----------
-    type :  cls(:class:`compas_timber.connections.Joint`)
-        The type of the joint.
-    kwargs : dict
-        The keyword arguments to be passed to the joint.
-
-    Attributes
-    ----------
-    type :  cls(:class:`compas_timber.connections.Joint`)
-        The type of the joint.
-    kwargs : dict
-        The keyword arguments to be passed to the joint.
-
-    """
-
-    def __init__(self, type, **kwargs):
-        self.type = type
-        self.kwargs = kwargs
-
-    def __repr__(self):
-        return "{}({}{})".format(JointOptions.__name__, self.type, self.kwargs)
-
-    def ToString(self):
-        return repr(self)
-
-    def is_identical(self, other):
-        return isinstance(other, JointOptions) and self.kwargs == other.kwargs
