+++ src/compas_timber/ghpython/__init__.py
@@ -3,6 +3,7 @@ from .workflow import CategoryRule
 from .workflow import TopologyRule
 from .workflow import DirectRule
 from .workflow import FeatureDefinition
+from .workflow import JointOptions
 
 
 __all__ = [
@@ -11,4 +12,5 @@ __all__ = [
     "TopologyRule",
     "DirectRule",
     "FeatureDefinition",
+    "JointOptions"
 ]
