+++ src/compas_timber/ghpython/components/CT_Joint_Options_LButt/code.py
@@ -2,7 +2,7 @@ from ghpythonlib.componentbase import executingcomponent as component
 
 
 from compas_timber.connections import LButtJoint
-from compas_timber.connections.joint import JointOptions
+from compas_timber.ghpython import JointOptions
 
 
 class MyComponent(component):
